ref: moved drag-drop components into /drag-drop
